refactor: reorganize wizard components
