+++ app/(wizard)/dashboard/new/components/dialogs/pitch-confirmation-dialog.tsx
