+++ app/(wizard)/dashboard/new/components/index.ts
@@ -1,0 +1,15 @@
+/**
+ * Barrel file for pitch wizard components
+ */
+
+export { default as PitchWizard } from "./wizard"
+export { default as WizardHeader } from "./progress/wizard-header"
+export { default as SectionProgressBar } from "./progress/section-progress-bar"
+export { default as MobileProgressHeader } from "./progress/mobile-progress-header"
+export { default as PitchConfirmationDialog } from "./dialogs/pitch-confirmation-dialog"
+export { default as StarFieldComponent } from "./utilities/star-field-component"
+export { default as WordCountIndicator } from "./utilities/word-count-indicator"
+export { default as AIThinkingLoader } from "./utilities/ai-thinking-loader"
+export { default as CheckStoredPitch } from "./utilities/check-stored-pitch"
+
+export * from "./wizard/schema"
