+++ app/(wizard)/dashboard/new/components/progress/mobile-progress-header.tsx
