+++ app/(wizard)/dashboard/new/components/progress/section-progress-bar.tsx
