+++ app/(wizard)/dashboard/new/components/progress/wizard-header.tsx
