+++ app/(wizard)/dashboard/new/components/steps/action-step.tsx
@@ -6,8 +6,8 @@ import {
   PitchWizardFormData,
   starExampleSchema,
   actionStepSchema
-} from "./pitch-wizard/schema"
-import WordCountIndicator from "./word-count-indicator"
+} from "../wizard/schema"
+import WordCountIndicator from "../utilities/word-count-indicator"
 import { useState, useEffect } from "react"
 import { v4 as uuidv4 } from "uuid"
 import {
