+++ app/(wizard)/dashboard/new/components/steps/result-step.tsx
@@ -2,8 +2,8 @@
 
 // Result step collects the final outcome and its benefits
 import { useFormContext } from "react-hook-form"
-import { PitchWizardFormData, starExampleSchema } from "./pitch-wizard/schema"
-import StarFieldComponent from "./star-field-component"
+import { PitchWizardFormData, starExampleSchema } from "../wizard/schema"
+import StarFieldComponent from "../utilities/star-field-component"
 
 interface ResultStepProps {
   /**
