+++ app/(wizard)/dashboard/new/components/steps/situation-step.tsx
@@ -2,8 +2,8 @@
 
 // Collects information about the Situation portion of a STAR example
 import { useFormContext } from "react-hook-form"
-import { PitchWizardFormData, starExampleSchema } from "./pitch-wizard/schema"
-import StarFieldComponent from "./star-field-component"
+import { PitchWizardFormData, starExampleSchema } from "../wizard/schema"
+import StarFieldComponent from "../utilities/star-field-component"
 
 interface SituationStepProps {
   /**
