+++ app/(wizard)/dashboard/new/components/utilities/ai-thinking-loader.tsx
