+++ app/(wizard)/dashboard/new/components/utilities/check-stored-pitch.tsx
