+++ app/(wizard)/dashboard/new/components/utilities/word-count-indicator.tsx
