+++ app/(wizard)/dashboard/new/components/wizard/index.tsx
