+++ app/(wizard)/dashboard/new/components/wizard/schema.ts
